+++ gax/test/browser-test/test/test.grpc-fallback.ts
@@ -93,8 +93,8 @@ describe('createStub', () => {
     assert(echoStub.pagedExpand instanceof Function);
     assert(echoStub.wait instanceof Function);
 
-    // There should be 10 methods for the echo service
-    assert.strictEqual(Object.keys(echoStub).length, 10);
+    // There should be 11 methods for the echo service
+    assert.strictEqual(Object.keys(echoStub).length, 11);
 
     // Each of the service methods should take 4 arguments (so that it works
     // with createApiCall)
@@ -109,8 +109,8 @@ describe('createStub', () => {
     assert(echoStub.collect instanceof Function);
     assert(echoStub.chat instanceof Function);
 
-    // There should be 10 methods for the echo service
-    assert.strictEqual(Object.keys(echoStub).length, 10);
+    // There should be 11 methods for the echo service
+    assert.strictEqual(Object.keys(echoStub).length, 11);
 
     // Each of the service methods should take 4 arguments (so that it works
     // with createApiCall)