+++ tools/src/compileProtos.ts
@@ -21,6 +21,7 @@ import * as path from 'path';
 import * as util from 'util';
 import * as pbjs from 'protobufjs-cli/pbjs';
 import * as pbts from 'protobufjs-cli/pbts';
+import {walkUp} from 'walk-up-path';
 
 export const gaxProtos = path.join(
   require.resolve('google-gax'),
@@ -238,6 +239,13 @@ async function buildListOfProtos(
   return result;
 }
 
+interface CompileProtosOptions {
+  skipJson?: boolean;
+  esm?: boolean;
+  keepCase?: boolean;
+  forceNumber?: boolean;
+}
+
 /**
  * Runs `pbjs` to compile the given proto files, placing the result into
  * `./protos/protos.json`. No support for changing output filename for now
@@ -249,10 +257,16 @@ async function buildListOfProtos(
 async function compileProtos(
   rootName: string,
   protos: string[],
-  skipJson = false,
-  esm = false
+  options: CompileProtosOptions
 ): Promise<void> {
-  if (!skipJson) {
+  const extraArgs = [];
+  if (options.keepCase) {
+    extraArgs.push('--keep-case');
+  }
+  if (options.forceNumber) {
+    extraArgs.push('--force-number');
+  }
+  if (!options.skipJson) {
     // generate protos.json file from proto list
     const jsonOutput = path.join('protos', 'protos.json');
     if (protos.length === 0) {
@@ -263,6 +277,7 @@ async function compileProtos(
     const pbjsArgs4JSON = [
       '--target',
       'json',
+      ...extraArgs,
       '-p',
       'protos',
       '-p',
@@ -275,7 +290,7 @@ async function compileProtos(
   }
 
   // generate protos/protos.js from protos.json
-  const jsOutput = esm
+  const jsOutput = options.esm
     ? path.join('protos', 'protos.cjs')
     : path.join('protos', 'protos.js');
   const pbjsArgs4js = [
@@ -283,6 +298,7 @@ async function compileProtos(
     rootName,
     '--target',
     'static-module',
+    ...extraArgs,
     '-p',
     'protos',
     '-p',
@@ -298,13 +314,14 @@ async function compileProtos(
   await writeFile(jsOutput, jsResult);
 
   let jsOutputEsm;
-  if (esm) {
+  if (options.esm) {
     jsOutputEsm = path.join('protos', 'protos.js');
     const pbjsArgs4jsEsm = [
       '-r',
       rootName,
       '--target',
       'static-module',
+      ...extraArgs,
       '-p',
       'protos',
       '-p',
@@ -324,7 +341,7 @@ async function compileProtos(
 
   // generate protos/protos.d.ts
   const tsOutput = path.join('protos', 'protos.d.ts');
-  const pbjsArgs4ts = [esm ? jsOutputEsm! : jsOutput, '-o', tsOutput];
+  const pbjsArgs4ts = [options.esm ? jsOutputEsm! : jsOutput, '-o', tsOutput];
   await pbtsMain(pbjsArgs4ts);
 
   let tsResult = (await readFile(tsOutput)).toString();
@@ -344,14 +361,16 @@ export async function generateRootName(directories: string[]): Promise<string> {
   // It's OK to play some guessing game here: if we locate `package.json`
   // with a package name, we'll use it; otherwise, we'll fallback to 'default'.
   for (const directory of directories) {
-    const packageJson = path.resolve(directory, '..', 'package.json');
-    if (fs.existsSync(packageJson)) {
-      const json = JSON.parse((await readFile(packageJson)).toString()) as {
-        name: string;
-      };
-      const name = json.name.replace(/[^\w\d]/g, '_');
-      const hopefullyUniqueName = `${name}_protos`;
-      return hopefullyUniqueName;
+    for (const p of walkUp(path.resolve(directory, '..'))) {
+      const packageJson = path.join(p, 'package.json');
+      if (fs.existsSync(packageJson)) {
+        const json = JSON.parse((await readFile(packageJson)).toString()) as {
+          name: string;
+        };
+        const name = json.name.replace(/[^\w\d]/g, '_');
+        const hopefullyUniqueName = `${name}_protos`;
+        return hopefullyUniqueName;
+      }
     }
   }
   return 'default';
@@ -372,6 +391,8 @@ export async function main(parameters: string[]): Promise<void> {
   const protoJsonFiles: string[] = [];
   let skipJson = false;
   let esm = false;
+  let keepCase = false;
+  let forceNumber = false;
   const directories: string[] = [];
   for (const parameter of parameters) {
     if (parameter === '--skip-json') {
@@ -382,6 +403,14 @@ export async function main(parameters: string[]): Promise<void> {
       esm = true;
       continue;
     }
+    if (parameter === '--keep-case') {
+      keepCase = true;
+      continue;
+    }
+    if (parameter === '--force-number') {
+      forceNumber = true;
+      continue;
+    }
     // it's not an option so it's a directory
     const directory = parameter;
     directories.push(directory);
@@ -390,10 +419,15 @@ export async function main(parameters: string[]): Promise<void> {
   const rootName = await generateRootName(directories);
   if (esm) {
     const esmProtos = await buildListOfProtos(protoJsonFiles, esm);
-    await compileProtos(rootName, esmProtos, skipJson, esm);
+    await compileProtos(rootName, esmProtos, {
+      skipJson,
+      esm,
+      keepCase,
+      forceNumber,
+    });
   }
   const protos = await buildListOfProtos(protoJsonFiles, esm);
-  await compileProtos(rootName, protos, skipJson, esm);
+  await compileProtos(rootName, protos, {skipJson, esm, keepCase, forceNumber});
 }
 
 /**
@@ -418,7 +452,7 @@ function usage() {
 if (require.main === module) {
   if (process.argv.length <= 2) {
     usage();
-    // eslint-disable-next-line no-process-exit
+    // eslint-disable-next-line n/no-process-exit
     process.exit(1);
   }
   // argv[0] is node.js binary, argv[1] is script path