+++ tools/test/compileProtos.ts
@@ -281,9 +281,16 @@ describe('compileProtos tool', () => {
     assert.strictEqual(rootName, '_org_fake_package_protos');
   });
 
-  it('falls back to the default name for protobuf root if unable to guess', async () => {
+  it('uses the nearest package.json to guess the root name', async () => {
     const rootName = await compileProtos.generateRootName([
       path.join(__dirname, 'protoLists', 'empty'),
+    ]);
+    assert.strictEqual(rootName, 'gapic_tools_protos');
+  });
+
+  it('falls back to the default name for protobuf root if unable to guess', async () => {
+    const rootName = await compileProtos.generateRootName([
+      '/nonexistent/empty',
     ]);
     assert.strictEqual(rootName, 'default');
   });
@@ -314,5 +321,49 @@ describe('compileProtos tool', () => {
       assert(ts.toString().includes(reformate));
       assert.equal(js.toString().includes(link), false);
     }
+  });
+
+  it('converts names to camelCase and uses Long by default', async function () {
+    this.timeout(20000);
+    const dirName = path.join(testDir, 'protoLists', 'parameters');
+    await compileProtos.main([dirName]);
+    const jsonBuf = await readFile(expectedJsonResultFile);
+    const json = JSON.parse(jsonBuf.toString());
+    const js = await readFile(expectedJSResultFile);
+    assert(json.nested.test.nested.Test.fields.snakeCaseField);
+    assert(js.includes('@property {number|Long|null} [checkForceNumber]'));
+  });
+
+  it('understands --keep-case', async function () {
+    this.timeout(20000);
+    const dirName = path.join(testDir, 'protoLists', 'parameters');
+    await compileProtos.main(['--keep-case', dirName]);
+    const jsonBuf = await readFile(expectedJsonResultFile);
+    const json = JSON.parse(jsonBuf.toString());
+    const js = await readFile(expectedJSResultFile);
+    assert(json.nested.test.nested.Test.fields.snake_case_field);
+    assert(js.includes('@property {number|Long|null} [check_force_number]'));
+  });
+
+  it('understands --force-number', async function () {
+    this.timeout(20000);
+    const dirName = path.join(testDir, 'protoLists', 'parameters');
+    await compileProtos.main(['--force-number', dirName]);
+    const jsonBuf = await readFile(expectedJsonResultFile);
+    const json = JSON.parse(jsonBuf.toString());
+    const js = await readFile(expectedJSResultFile);
+    assert(json.nested.test.nested.Test.fields.snakeCaseField);
+    assert(js.includes('@property {number|null} [checkForceNumber]'));
+  });
+
+  it('understands both --keep-case and --force-number', async function () {
+    this.timeout(20000);
+    const dirName = path.join(testDir, 'protoLists', 'parameters');
+    await compileProtos.main(['--keep-case', '--force-number', dirName]);
+    const jsonBuf = await readFile(expectedJsonResultFile);
+    const json = JSON.parse(jsonBuf.toString());
+    const js = await readFile(expectedJSResultFile);
+    assert(json.nested.test.nested.Test.fields.snake_case_field);
+    assert(js.includes('@property {number|null} [check_force_number]'));
   });
 });